renamed CTA link to ui link and added className prop for color support.
