--- src/client/core/components/CTALink/CTALink.scss
@@ -1,9 +1,0 @@
-@import '~styles/config/_colors.scss';
-
-.CTALink {
-  &__container {
-    a {
-      color: $black;
-    }
-  }
-}
