+++ src/client/core/components/UILink/UILink-test.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { MemoryRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import { mount } from 'enzyme';
+import thunk from 'redux-thunk';
+import configureStore from 'redux-mock-store';
+import initialState from '@mocks/state.json';
+import UILink from './index';
+
+describe('Component: CTALink', () => {
+  let wrapper;
+  let store;
+  const middlewares = [thunk];
+  const mockStore = configureStore(middlewares);
+  beforeEach(() => {
+    store = mockStore(initialState);
+    wrapper = mount(
+      <Provider store={store}>
+        <MemoryRouter>
+          <UILink />
+        </MemoryRouter>
+      </Provider>
+    );
+  });
+  it('renders without crashing', () => {
+    expect(wrapper.exists()).toBe(true);
+  });
+});
